--- leader-election-impl/src/main/java/com/palantir/paxos/history/models/PaxosHistoryOnSingleNode.java
@@ -1,30 +1,0 @@
-/*
- * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.paxos.history.models;
-
-import java.util.Map;
-
-import org.immutables.value.Value;
-
-import com.palantir.paxos.NamespaceAndUseCase;
-
-@Value.Immutable
-public interface PaxosHistoryOnSingleNode {
-
-    @Value.Parameter
-    Map<NamespaceAndUseCase, LearnerAndAcceptorRecords> history();
-}
