+++ settings.gradle
@@ -64,6 +64,7 @@ include ":timelock-api:timelock-api-dialogue"
 include ":timelock-api:timelock-api-objects"
 include ":timelock-corruption-detection"
 include ":timelock-impl"
+include ":timelock-corruption"
 include ":timelock-server"
 include ":timelock-server-benchmark-client"
 include ":timelock-server-benchmark-cluster"
