+++ timelock-corruption/build.gradle
@@ -1,0 +1,17 @@
+apply from: "../gradle/shared.gradle"
+
+dependencies {
+    compile project(":leader-election-impl")
+
+    compile group: 'com.google.guava', name: 'guava'
+    compile group: 'com.palantir.conjure.java.api', name: 'service-config'
+
+    annotationProcessor group: 'org.immutables', name: 'value'
+    compileOnly 'org.immutables:value::annotations'
+
+    testCompile group: 'org.assertj', name: 'assertj-core'
+    testCompile (group: 'org.mockito', name: 'mockito-core') {
+        // Mockito version doesn't agree with our JUnit version
+        exclude group: 'org.hamcrest', module: 'hamcrest-core'
+    }
+}
